requirements.txt
flask==3.0.0
torch==2.2.0
torchvision==0.17.0
tensorflow-cpu==2.15.0
pillow==10.0.0
numpy==1.24.0
gunicorn==21.2.0